% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI.R
\name{BMI}
\alias{BMI}
\title{Body Mass Index}
\usage{
BMI(height, weight, PRINT = FALSE)
}
\arguments{
\item{height}{Input height (in cm)}

\item{weight}{Input weight (in kg)}

\item{PRINT}{if print is TRUE, then print level (Default is FALSE)}
}
\value{
BMI
}
\description{
Caculate BMI
}
